+++ tests/Instances/Data/Functor/Trans.hs
@@ -1,3 +1,4 @@
+{-# LANGUAGE CPP                        #-}
 {-# LANGUAGE FlexibleContexts           #-}
 {-# LANGUAGE GeneralizedNewtypeDeriving #-}
 {-# LANGUAGE StandaloneDeriving         #-}
@@ -15,7 +16,9 @@ Provides 'Arbitrary' instances for functor transformers.
 module Instances.Data.Functor.Trans () where
 
 import Data.Functor.Compose  (Compose(..))
+#if !(MIN_VERSION_QuickCheck(2,9,0))
 import Data.Functor.Constant (Constant(..))
+#endif
 import Data.Functor.Product  (Product(..))
 import Data.Functor.Reverse  (Reverse(..))
 import Data.Functor.Sum      (Sum(..))
@@ -26,7 +29,9 @@ import Prelude.Compat
 import Test.QuickCheck (Arbitrary(..), oneof)
 
 deriving instance Arbitrary (f (g a)) => Arbitrary (Compose f g a)
+#if !(MIN_VERSION_QuickCheck(2,9,0))
 deriving instance Arbitrary a         => Arbitrary (Constant a b)
+#endif
 deriving instance Arbitrary (f a)     => Arbitrary (Reverse f a)
 
 instance (Arbitrary (f a), Arbitrary (g a)) => Arbitrary (Product f g a) where